+++ model_utils/fields.py
@@ -70,6 +70,11 @@ class StatusField(models.CharField):
         self._choices = [(0, 'dummy')]
         super(StatusField, self).contribute_to_class(cls, name)
 
+    def deconstruct(self):
+        name, path, args, kwargs = super(StatusField, self).deconstruct()
+        kwargs['no_check_for_status'] = True
+        return name, path, args, kwargs
+
 
 class MonitorField(models.DateTimeField):
     """
@@ -112,6 +117,14 @@ class MonitorField(models.DateTimeField):
                 setattr(model_instance, self.attname, value)
                 self._save_initial(model_instance.__class__, model_instance)
         return super(MonitorField, self).pre_save(model_instance, add)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(MonitorField, self).deconstruct()
+        if self.monitor is not None:
+            kwargs['monitor'] = self.monitor
+        if self.when is not None:
+            kwargs['when'] = self.when
+        return name, path, args, kwargs
 
 
 SPLIT_MARKER = getattr(settings, 'SPLIT_MARKER', '<!-- split -->')